Make accented characters saved properly (utf8)
* Update JDBC URL samples in settings.sample.xml with unicode and UTF-8.
* update JDBC URL in application.properties in main and test with
  unicode and UTF-8.
* Update README with comments on UTF-8 usage in MySQL/MariaDB.

Basically `utf8mb4` encoding should be used, because `utf8` is broken,
however if one uses this by default there's a problem with key lenghts
or similar. This can be solved by enabling large keys ... however that's
problematic, because in older versions of MySQL/MariabDB versions this
cannot be configured via a global system property, only by explicitly in
table declarations, but that's ... also problematic. So, for now `utf8`
is used by this project, because "it's good enough". The project thus
can be used with databases using `utf8` and `utf8mb4` encoding as well.